Move and rename files for more consistant orgainization
