+++ src/navigation/ElectionsStack.js
@@ -1,6 +1,6 @@
 import { createStackNavigator } from 'react-navigation';
 import ElectionsScreen from '../screens/ElectionsScreen';
-import CandidateDetailScreen from '../screens/CandidateDetailScreen';
+import CandidateDetailScreen from '../screens/CandidateDetail';
 
 import styles from './styles';
 import colors from '../styles/colors';
