+++ src/screens/CandidateDetail/CandidateDetail.js
